+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 * Added
+  * Added new test for Terra Slider component for long field labels.
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
-
+  
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.
   * Removed alpha notice for FlowsheetDataGrid.